update model repository to have all the model logic
