+++ nlp.services/NlpRepository.cs
@@ -19,17 +19,17 @@ namespace nlp.services
     {
         private readonly ILogger<NlpRepository<T>> _logger;
         private readonly IMemoryCache _cache;
+        private readonly IModelRepository<T> _modelrepo;
         private readonly ITextMiningRepository<T> _txtrepo;
-        private readonly Models<T> _models;
 
         private ICollection<ICategory> _categories { get; set; } = new List<ICategory>();
 
-        public NlpRepository(ILogger<NlpRepository<T>> logger, IMemoryCache cache, ITextMiningRepository<T> txtrepo, Models<T> models)
+        public NlpRepository(ILogger<NlpRepository<T>> logger, IMemoryCache cache, IModelRepository<T> modelrepo, ITextMiningRepository<T> txtrepo)
         {
             _logger = logger ?? throw new NlpException(HttpStatusCode.InternalServerError, nameof(logger));
             _cache = cache ?? throw new NlpException(HttpStatusCode.InternalServerError, nameof(cache));
+            _modelrepo = modelrepo ?? throw new NlpException(HttpStatusCode.InternalServerError, nameof(modelrepo));
             _txtrepo = txtrepo ?? throw new NlpException(HttpStatusCode.InternalServerError, nameof(txtrepo));
-            _models = models ?? throw new NlpException(HttpStatusCode.InternalServerError, nameof(models));
         }
 
         public INlpResponse Categorize(INlpRequest<T> Request, string Id = null, bool Summarize = false)
@@ -42,7 +42,7 @@ namespace nlp.services
 
             var models = new Stack<T>(new List<T>() { modelSettings.Model });
             var delimiters = modelSettings.Delimiters
-                .Union(_models.DefaultDelimiters)
+                .Union(_modelrepo._models.DefaultDelimiters)
                 .ToArray();
             var sw = new Stopwatch();
 
@@ -96,7 +96,7 @@ namespace nlp.services
         public IModelSettings<T> Parse(INlpRequest<T> Request, string Id = null)
         {
             if (Id != null)
-                return GetModelSettingsByModelId(Id);
+                return _modelrepo.GetModelSettingsByModelId(Id);
 
             if (Request == null)
                 throw new NlpException(HttpStatusCode.BadRequest,
@@ -117,7 +117,7 @@ namespace nlp.services
                     Delimiters = Request.Delimiters.Select(char.Parse).ToArray(),
                     Model = _cache.GetOrCreate(Request.Model.PublicKey, e =>
                         {
-                            e.SlidingExpiration = TimeSpan.FromSeconds(_models.TenMinutesCacheTimeSpan);
+                            e.SlidingExpiration = TimeSpan.FromSeconds(_modelrepo._models.TenMinutesCacheTimeSpan);
 
                             return Request.Model;
                         })
@@ -134,7 +134,7 @@ namespace nlp.services
                     Delimiters = Request.Delimiters.Select(char.Parse).ToArray(),
                     Model = _cache.GetOrCreate(Request.ModelId, e =>
                         {
-                            e.SlidingExpiration = TimeSpan.FromSeconds(_models.TenMinutesCacheTimeSpan);
+                            e.SlidingExpiration = TimeSpan.FromSeconds(_modelrepo._models.TenMinutesCacheTimeSpan);
 
                             return new T()
                             {
@@ -193,138 +193,7 @@ namespace nlp.services
                     break;
                 }
             }
-        }
-
-        public IEnumerable<T> GetModels()
-        {
-            return _cache.GetOrCreate(_models.All, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                return _models.All;
-            });
-        }
-        public IModel<T> GetModel(string Id)
-        {
-            var model = _cache.GetOrCreate(Id, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                return _models.All
-                    .FirstOrDefault(x => x.Id == Id);
-            });
-
-            if (model == null)
-            {
-                Guid.TryParse(Id, out Guid PublicKey);
-
-                return GetModel(PublicKey);
-            }
-
-            return model;
-        }
-        public IModel<T> GetModel(Guid PublicKey)
-        {
-            return _cache.GetOrCreate(PublicKey, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                return _models.All
-                    .FirstOrDefault(x => x.PublicKey == PublicKey);
-            });
-        }
-        public IModel<T> GetAnyModel(string Id)
-        {
-            return _cache.GetOrCreate(Id, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                var models = new Stack<T>(_models.All);
-
-                while (models.Any())
-                {
-                    var model = models.Pop() as IModel<T>;
-
-                    if (model.Id == Id) return model;
-
-                    if (model.Children.Any())
-                        model.Children
-                            .ToList()
-                            .ForEach(x =>
-                            {
-                                models.Push(x);
-                            });
-                }
-
-                return null;
-            });
-        }
-        public IEnumerable<IModelSettings<T>> GetModelsSettings()
-        {
-            return _cache.GetOrCreate(_models.Settings, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                return _models.Settings;
-            });
-        }
-        public IModelSettings<T> GetModelSettings(string Id)
-        {
-            return _cache.GetOrCreate(Id, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                return _models.Settings
-                    .FirstOrDefault(x => x.Id == Id);
-            });
-        }
-        public IModelSettings<T> GetModelSettingsByModelId(string Id)
-        {
-            return _cache.GetOrCreate(Id, e =>
-            {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
-
-                return _models.Settings
-                    .FirstOrDefault(x => x.Model.Id == Id);
-            });
-        }
-
-        public T AddModel(dynamic Request)
-        {
-            var jsonRequest = (JsonElement)Request;
-            var model = jsonRequest.GetProperty("model")
-                .GetRawText()
-                .DeserializeSelfReferencing<T>();
-
-            _cache.Set(model.PublicKey, model, DateTimeOffset.Now.AddSeconds(_models.DefaultCacheTimeSpan));
-
-            return model;
-        }
-        public T AddModel(INlpRequest<T> Request)
-        {
-            _cache.Set(Request.Model.PublicKey, Request.Model, DateTimeOffset.Now.AddSeconds(_models.DefaultCacheTimeSpan));
-
-            return Request.Model;
-        }
-        public T AddModel(T Model)
-        {
-            _cache.Set(Model.PublicKey, Model, DateTimeOffset.Now.AddSeconds(_models.DefaultCacheTimeSpan));
-
-            return Model;
-        }
-
-        public void DeleteModel(string Id)
-        {
-            var model = GetModel(Id);
-
-            try {  _cache.Remove(model?.Id); } catch (Exception) { throw new NlpException(HttpStatusCode.BadRequest, $"couldn't delete model with id={Id}"); }
-            try { _cache.Remove(model?.PublicKey); } catch (Exception) { throw new NlpException(HttpStatusCode.BadRequest, $"couldn't delete model with public key={model?.PublicKey}"); }
-        }
-        public void DeleteModel(string Id, Guid PublicKey)
-        {
-            try { _cache.Remove(Id); } catch (Exception) { throw new NlpException(HttpStatusCode.BadRequest, $"couldn't delete model with id={Id}"); }
-            try { _cache.Remove(PublicKey); } catch (Exception) { throw new NlpException(HttpStatusCode.BadRequest, $"couldn't delete model with public key={PublicKey}"); }
-        }
+        }      
 
         public object CategorizeSample()
         {
@@ -336,13 +205,13 @@ namespace nlp.services
 
             return _cache.GetOrCreate(requestContent, e =>
             {
-                e.SlidingExpiration = TimeSpan.FromSeconds(_models.DefaultCacheTimeSpan);
+                e.SlidingExpiration = TimeSpan.FromSeconds(_modelrepo._models.DefaultCacheTimeSpan);
 
                 return new
                 {
                     content = requestContent,
-                    categorization = Categorize(nlpRequest, _models.Financial.Id),
-                    model = _models.Financial
+                    categorization = Categorize(nlpRequest, _modelrepo._models.Financial.Id),
+                    model = _modelrepo._models.Financial
                 };
             });
         }
